chore: complete shared Go lint migration
Keep package-level Testify calls available so adopting the shared policy
does not require assertion-object churn. Preserve exact JSON comparisons
and use the kit ownership analysis for rows returned to callers.

Thread cancellation through storage and sync work, synchronize concurrent
tests on observable events, and preserve terminal lifetime beyond HTTP
requests. Retain documented interface and fallback exceptions.

+++ internal/parser/codex_cursor.go
@@ -122,7 +122,7 @@ func (s *codexCursorState) MarshalBinary() ([]byte, error) {
 	if err := writeStr(s.lastTaskEvent); err != nil {
 		return nil, err
 	}
-	if err := write(uint8(s.pendingCallCount)); err != nil {
+	if err := write(s.pendingCallCount); err != nil {
 		return nil, err
 	}
 	for i := range s.pendingCallCount {
