+++ internal/parser/file_identity_unix.go
@@ -9,7 +9,7 @@ import (
 
 func sourceFileIdentity(info os.FileInfo) (inode, device uint64) {
 	if stat, ok := info.Sys().(*syscall.Stat_t); ok {
-		return uint64(stat.Ino), uint64(stat.Dev)
+		return stat.Ino, uint64(stat.Dev)
 	}
 	return 0, 0
 }
