+++ internal/rawcapture/identity_unix.go
@@ -13,5 +13,5 @@ func stableFileIdentity(_ *os.File, info os.FileInfo) string {
 	if !ok {
 		return ""
 	}
-	return fmt.Sprintf("%d:%d", uint64(stat.Dev), uint64(stat.Ino))
+	return fmt.Sprintf("%d:%d", uint64(stat.Dev), stat.Ino)
 }
